requirements.txt
Django==4.0.2
django-allauth==0.48.0
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
python-dotenv==0.19.2
python-json-logger==2.0.2
psycopg2==2.9.3


